Move the inaction exercises under 'sandbox' package
