--- src/main/kotlin/inaction/ExampleRunner.kt
@@ -1,15 +1,0 @@
-package inaction
-
-// import inaction.chap02.runChap02
-// import inaction.chap03.runChap03
-// import inaction.chap04.runChap04
-// import inaction.chap05.runChap05
-// import inaction.chap06.runChap06
-// import inaction.chap07.runChap07
-// import inaction.chap08.runChap08
-import inaction.chap09.runChap09
-
-fun runExamples() {
-    // runChap02()
-    runChap09()
-}
