--- src/main/kotlin/inaction/chap03/strings/Util.kt
@@ -1,19 +1,0 @@
-package strings
-
-fun String.lastChar(): Char = get(length - 1)
-
-fun <T> Collection<T>.joinToString(
-    separator: String = ", ",
-    prefix: String = "",
-    postfix: String = ""
-): String {
-    val result = StringBuilder(prefix)
-
-    for ((index, element) in this.withIndex()) {
-        if (index > 0) result.append(separator)
-        result.append(element)
-    }
-
-    result.append(postfix)
-    return result.toString()
-}
