+++ src/main/kotlin/sandbox/inaction/chap02/learning/iterators.kt
@@ -1,0 +1,15 @@
+package sandbox.inaction.chap02.learning
+
+// Iterators and ranges
+fun fizzBuzz(i: Int) = when {
+    i % 15 == 0 -> "FizzBuzz "
+    i % 3 == 0 -> "Fizz "
+    i % 5 == 0 -> "Buzz "
+    else -> "$i "
+}
+
+fun recognize(c: Char) = when (c) {
+    in '0'..'9' -> "It's a digit!"
+    in 'a'..'z', in 'A'..'Z' -> "It's a letter!"
+    else -> "I don't know..."
+}
